requirements.txt
APScheduler==3.10.4
cryptography==41.0.3
cryptography==42.0.5
fastapi==0.111.0
numexpr==2.8.5
numexpr==2.10.0
pandas==2.2.2
psycopg2==2.9.9
psycopg2_binary==2.9.9
pydantic==2.7.2
python-dotenv==1.0.1
redis==5.0.3
SQLAlchemy==1.4.52
SQLAlchemy==2.0.30
streamlit==1.26.0
streamlit==1.33.0
tradingview_datafeed==2.1.1
tvdatafeed==2.1.0
uvicorn==0.30.0
